requirements.txt
boto3
ask-sdk-core
openai
python-dotenv
pandas
hermetrics
arduino-iot-client